WP Security Update

+++ wp-content/plugins/wpforms/src/Pro/Admin/Entries/Export/File.php
@@ -2,17 +2,13 @@
 
 namespace WPForms\Pro\Admin\Entries\Export;
 
-use \Goodby\CSV\Export\Standard\Exporter;
-use \Goodby\CSV\Export\Standard\ExporterConfig;
+use Goodby\CSV\Export\Standard\CsvFileObject;
+use WPForms\Helpers\Transient;
 
 /**
  * File-related routines.
  *
- * @since      1.5.5
- * @author     WPForms
- * @package    WPForms\Pro\Admin\Entries\Export
- * @license    GPL-2.0+
- * @copyright  Copyright (c) 2019, WPForms LLC
+ * @since 1.5.5
  */
 class File {
 
@@ -46,9 +42,9 @@ class File {
 	 */
 	public function hooks() {
 
-		add_action( 'wpforms_tools_init', array( $this, 'entries_export_download_file' ) );
-		add_action( 'wpforms_tools_init', array( $this, 'single_entry_export_download_file' ) );
-		add_action( 'admin_init', array( $this, 'remove_old_export_files' ) );
+		add_action( 'wpforms_tools_init', [ $this, 'entries_export_download_file' ] );
+		add_action( 'wpforms_tools_init', [ $this, 'single_entry_export_download_file' ] );
+		add_action( Export::TASK_CLEANUP, [ $this, 'remove_old_export_files' ] );
 	}
 
 	/**
@@ -56,21 +52,78 @@ class File {
 	 *
 	 * @since 1.5.5
 	 *
-	 * @param array $export_data  Export data array.
 	 * @param array $request_data Request data array.
 	 */
-	public function write_csv( $export_data, $request_data ) {
+	public function write_csv( $request_data ) {
 
 		$export_file = $this->get_tmpfname( $request_data );
 
-		// Include Exporter.
-		require_once WPFORMS_PLUGIN_DIR . 'vendor/autoload.php';
+		if ( empty( $export_file ) ) {
+			return;
+		}
 
-		$config = new ExporterConfig();
-		$config->setDelimiter( $this->export->configuration['csv_export_separator'] );
-		$config->setFileMode( 'a' );
-		$exporter = new Exporter( $config );
-		$exporter->export( $export_file, $export_data );
+		$csv       = new CsvFileObject( $export_file, 'a' );
+		$enclosure = '"';
+
+		$csv->fputcsv( $request_data['columns_row'], $this->export->configuration['csv_export_separator'], $enclosure );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->get( 'entry' )->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$csv->fputcsv( $entry, $this->export->configuration['csv_export_separator'], $enclosure );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$csv->fflush();
+	}
+
+	/**
+	 * Write data to .xlsx file.
+	 *
+	 * @since 1.6.5
+	 *
+	 * @param array $request_data Request data array.
+	 */
+	public function write_xlsx( $request_data ) {
+
+		$export_file = $this->get_tmpfname( $request_data );
+
+		if ( empty( $export_file ) ) {
+			return;
+		}
+
+		$writer = new \XLSXWriter();
+
+		$writer->setTempDir( $this->get_tmpdir() );
+
+		$sheet_name = 'WPForms';
+
+		$widths = [];
+
+		foreach ( $request_data['columns_row'] as $header ) {
+			$widths[] = strlen( $header ) + 2;
+		}
+
+		$writer->writeSheetHeader( $sheet_name, array_fill_keys( $request_data['columns_row'], 'string' ), [ 'widths' => $widths ] );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->get( 'entry' )->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$writer->writeSheetRow( $sheet_name, $entry, [ 'wrap_text' => true ] );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$writer->writeToFile( $export_file );
 	}
 
 	/**
@@ -82,24 +135,25 @@ class File {
 	 */
 	public function get_tmpdir() {
 
-		$uploads              = wp_upload_dir();
-		$wpforms_uploads_root = trailingslashit( $uploads['basedir'] ) . 'wpforms';
+		$upload_dir  = wpforms_upload_dir();
+		$upload_path = $upload_dir['path'];
 
-		// Apply filter to allow redefine tmp directory.
-		$custom_uploads_root = apply_filters( 'wpforms_upload_root', $wpforms_uploads_root );
-		if ( wp_is_writable( $custom_uploads_root ) ) {
-			$wpforms_uploads_root = $custom_uploads_root;
+		$export_path = trailingslashit( $upload_path ) . 'export';
+		if ( ! file_exists( $export_path ) ) {
+			wp_mkdir_p( $export_path );
 		}
 
-		$export_dir = trailingslashit( $wpforms_uploads_root ) . 'export';
-		if ( ! file_exists( $export_dir ) ) {
-			wp_mkdir_p( $export_dir );
-		}
+		// Check if the .htaccess exists in the upload directory, if not - create it.
+		wpforms_create_upload_dir_htaccess_file();
+
+		// Check if the index.html exists in the directories, if not - create it.
+		wpforms_create_index_html_file( $upload_path );
+		wpforms_create_index_html_file( $export_path );
 
 		// Normalize slashes for Windows.
-		$export_dir = wp_normalize_path( $export_dir );
+		$export_path = wp_normalize_path( $export_path );
 
-		return $export_dir;
+		return $export_path;
 	}
 
 	/**
@@ -113,11 +167,77 @@ class File {
 	 */
 	public function get_tmpfname( $request_data ) {
 
+		if ( empty( $request_data ) ) {
+			return '';
+		}
+
 		$export_dir  = $this->get_tmpdir();
 		$export_file = $export_dir . '/' . sanitize_key( $request_data['request_id'] );
 		touch( $export_file );
 
 		return $export_file;
+	}
+
+	/**
+	 * Send HTTP headers for .csv file download.
+	 *
+	 * @since 1.5.5.1
+	 *
+	 * @param string $file_name File name.
+	 */
+	public function http_headers( $file_name ) {
+
+		$file_name = empty( $file_name ) ? 'wpforms-entries.csv' : $file_name;
+
+		nocache_headers();
+		header( 'Content-Description: File Transfer' );
+		header( 'Content-Type: text/csv' );
+		header( 'Content-Disposition: attachment; filename=' . $file_name );
+		header( 'Content-Transfer-Encoding: binary' );
+	}
+
+	/**
+	 * Output the file.
+	 *
+	 * @since 1.6.0.2
+	 *
+	 * @param array $request_data Request data.
+	 *
+	 * @throws \Exception In case of file error.
+	 */
+	public function output_file( $request_data ) {
+
+		$export_file = $this->get_tmpfname( $request_data );
+
+		if ( empty( $export_file ) ) {
+			throw new \Exception( $this->export->errors['unknown_request'] );
+		}
+
+		clearstatcache( true, $export_file );
+
+		// Remove the transient.
+		Transient::delete( 'wpforms-tools-entries-export-request-' . $request_data['request_id'] );
+
+		if ( ! is_readable( $export_file ) || is_dir( $export_file ) ) {
+			throw new \Exception( $this->export->errors['file_not_readable'] );
+		}
+
+		if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
+			throw new \Exception( $this->export->errors['file_empty'] );
+		}
+
+		$entry_suffix = ! empty( $request_data['db_args']['entry_id'] ) ? '-entry-' . $request_data['db_args']['entry_id'] : '';
+
+		$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . $entry_suffix . '-' . current_time( 'Y-m-d-H-i-s' ) . '.' . $request_data['type'];
+
+		$this->http_headers( $file_name );
+
+		readfile( $export_file ); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_readfile
+
+		// Schedule clean up.
+		$this->schedule_remove_old_export_files();
+
+		exit;
 	}
 
 	/**
@@ -131,7 +251,7 @@ class File {
 
 		$args = $this->export->data['get_args'];
 
-		if ( 'wpforms_tools_entries_export_download' !== $args['action'] ) {
+		if ( $args['action'] !== 'wpforms_tools_entries_export_download' ) {
 			return;
 		}
 
@@ -140,7 +260,7 @@ class File {
 			// Security check.
 			if (
 				! wp_verify_nonce( $args['nonce'], 'wpforms-tools-entries-export-nonce' ) ||
-				! wpforms_current_user_can()
+				! wpforms_current_user_can( 'view_entries' )
 			) {
 				throw new \Exception( $this->export->errors['security'] );
 			}
@@ -151,27 +271,9 @@ class File {
 			}
 
 			// Get stored request data.
-			$request_data = get_transient( 'wpforms-tools-entries-export-request-' . $args['request_id'] );
+			$request_data = Transient::get( 'wpforms-tools-entries-export-request-' . $args['request_id'] );
 
-			$export_file = $this->get_tmpfname( $request_data );
-
-			clearstatcache( true, $export_file );
-
-			if ( ! is_readable( $export_file ) ) {
-				throw new \Exception( $this->export->errors['file_not_readable'] );
-			}
-
-			if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
-				throw new \Exception( $this->export->errors['file_empty'] );
-			}
-
-			$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . '-' . date( 'Y-m-d-H-i-s' ) . '.csv';
-			nocache_headers();
-			header( 'Content-Disposition: attachment; filename=' . $file_name );
-			header( 'Content-Transfer-Encoding: binary' );
-
-			readfile( $export_file ); // phpcs:ignore
-			exit;
+			$this->output_file( $request_data );
 
 		} catch ( \Exception $e ) {
 			// phpcs:disable
@@ -180,17 +282,18 @@ class File {
 				$error .= '<br><b>WPFORMS DEBUG</b>: ' . $e->__toString();
 			}
 			$error = str_replace( "'", '&#039;', $error );
+			$js_error = str_replace( "\n", '', $error );
 
 			echo "
 			<script>
 				( function() {
 					var w = window;
-					if ( w.frameElement != null &&
-						 w.frameElement.nodeName === 'IFRAME' &&
-						 w.parent.jQuery )
-					{
-						w.parent.jQuery( w.parent.document ).trigger( 'csv_file_error', [ '" . str_replace( "\n", '', $error ) . "' ] );
-						w.parent.WPFormsEntriesExport.displaySubmitSpinner( false );
+					if (
+						w.frameElement !== null &&
+						w.frameElement.nodeName === 'IFRAME' &&
+						w.parent.jQuery
+					) {
+						w.parent.jQuery( w.parent.document ).trigger( 'csv_file_error', [ '" . wp_slash( $js_error ) . "' ] );
 					}
 				} )();
 			</script>
@@ -230,7 +333,7 @@ class File {
 			// Security check.
 			if (
 				! wp_verify_nonce( $args['nonce'], 'wpforms-tools-single-entry-export-nonce' ) ||
-				! wpforms_current_user_can()
+				! wpforms_current_user_can( 'view_entries' )
 			) {
 				throw new \Exception( $this->export->errors['security'] );
 			}
@@ -240,31 +343,13 @@ class File {
 
 			$this->export->ajax->request_data = $request_data;
 
-			// Get export data.
-			$export_data = $this->export->ajax->get_data();
-
-			// Writing to csv file.
-			$this->write_csv( $export_data, $request_data );
-
-			$export_file = $this->get_tmpfname( $request_data );
-
-			clearstatcache( true, $export_file );
-
-			if ( ! is_readable( $export_file ) ) {
-				throw new \Exception( $this->export->errors['file_not_readable'] );
+			if ( empty( $request_data['type'] ) || $request_data['type'] === 'csv' ) {
+				$this->write_csv( $request_data );
+			} elseif ( $request_data['type'] === 'xlsx' ) {
+				$this->write_xlsx( $request_data );
 			}
 
-			if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
-				throw new \Exception( $this->export->errors['file_empty'] );
-			}
-
-			$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . '-entry-' . $request_data['db_args']['entry_id'] . '-' . date( 'Y-m-d-H-i-s' ) . '.csv';
-			nocache_headers();
-			header( 'Content-Disposition: attachment; filename=' . $file_name );
-			header( 'Content-Transfer-Encoding: binary' );
-
-			readfile( $export_file ); // phpcs:ignore
-			exit;
+			$this->output_file( $request_data );
 
 		} catch ( \Exception $e ) {
 
@@ -273,24 +358,38 @@ class File {
 				$error .= '<br><b>WPFORMS DEBUG</b>: ' . $e->__toString();
 			}
 
-			\WPForms_Admin_Notice::error( $error );
+			\WPForms\Admin\Notice::error( $error );
 		}
 	}
 
 	/**
-	 * Garbage clearing.
-	 * Actually we need to clear only temporary files
-	 * because transients clears automagically.
+	 * Register a cleanup task to remove temp export files.
 	 *
-	 * TODO: rewrite to use wp-cron or actions scheduler.
+	 * @since 1.6.5
+	 */
+	public function schedule_remove_old_export_files() {
+
+		$tasks = wpforms()->get( 'tasks' );
+
+		if ( ! empty( $tasks->is_scheduled( Export::TASK_CLEANUP ) ) ) {
+			return;
+		}
+
+		$tasks->create( Export::TASK_CLEANUP )
+		      ->recurring( time() + 60, $this->export->configuration['request_data_ttl'] )
+		      ->params()
+		      ->register();
+	}
+
+	/**
+	 * Clear temporary files.
 	 *
 	 * @since 1.5.5
+	 * @since 1.6.5 Stop using transients. Now we use ActionScheduler for this task.
 	 */
 	public function remove_old_export_files() {
 
-		if ( (bool) get_transient( 'wpforms_entries_export_tmpdata_cleared' ) ) {
-			return;
-		}
+		clearstatcache();
 
 		$files = glob( $this->get_tmpdir() . '/*' );
 		$now   = time();
@@ -298,12 +397,11 @@ class File {
 		foreach ( $files as $file ) {
 			if (
 				is_file( $file ) &&
+				pathinfo( $file, PATHINFO_BASENAME ) !== 'index.html' &&
 				( $now - filemtime( $file ) ) > $this->export->configuration['request_data_ttl']
 			) {
 				unlink( $file );
 			}
 		}
-
-		set_transient( 'wpforms_entries_export_tmpdata_cleared', 'yes', HOUR_IN_SECONDS );
 	}
 }
